update đường dẫn side links, gộp folder, đổi tên partList
